+++ Assets/YukimaruGames/Terminal/Runtime/Application/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b951f9343ab56ae4fb82fd788a529fd1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
