+++ Assets/YukimaruGames/Terminal/Runtime/Application/Mappers.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5a1323e57ab368c44baed1c6b6bb3c7b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
