+++ Assets/YukimaruGames/Terminal/Runtime/Application/Mappers/LogMapper.cs
@@ -1,7 +1,8 @@
 using System;
+using YukimaruGames.Terminal.Application.Models;
 using YukimaruGames.Terminal.Domain.Abstractions.Models;
 
-namespace YukimaruGames.Terminal.Application.Commands
+namespace YukimaruGames.Terminal.Application.Mappers
 {
     public static class LogMapper
     {
