+++ Assets/YukimaruGames/Terminal/Runtime/Application/Services/TerminalService.cs
@@ -1,13 +1,15 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using YukimaruGames.Terminal.Application.Commands;
+using YukimaruGames.Terminal.Application.Interfaces;
+using YukimaruGames.Terminal.Application.Mappers;
+using YukimaruGames.Terminal.Application.Models;
 using YukimaruGames.Terminal.Domain.Abstractions.Exceptions;
 using YukimaruGames.Terminal.Domain.Abstractions.Interfaces;
 using YukimaruGames.Terminal.Domain.Abstractions.Models;
 using YukimaruGames.Terminal.SharedKernel;
 
-namespace YukimaruGames.Terminal.Application.Core
+namespace YukimaruGames.Terminal.Application.Services
 {
     /// <summary>
     /// ドメインサービスを全般を統括し、外部へAPIを提供するアプリケーションサービス.
