+++ Assets/YukimaruGames/Terminal/Runtime/Application/Services/TerminalService.cs.meta
