+++ Assets/YukimaruGames/Terminal/Runtime/Runtime/Installer/TerminalStandardInstaller.cs
@@ -13,7 +13,8 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
-using YukimaruGames.Terminal.Application.Core;
+using YukimaruGames.Terminal.Application.Interfaces;
+using YukimaruGames.Terminal.Application.Services;
 using YukimaruGames.Terminal.Domain.Abstractions.Interfaces;
 using YukimaruGames.Terminal.Domain.Services;
 using YukimaruGames.Terminal.Infrastructure.Commands;
