+++ Assets/YukimaruGames/Terminal/Runtime/UI/Main/Coordinator.cs
@@ -1,6 +1,6 @@
 using System;
 using UnityEngine;
-using YukimaruGames.Terminal.Application.Core;
+using YukimaruGames.Terminal.Application.Interfaces;
 using YukimaruGames.Terminal.UI.Core;
 using YukimaruGames.Terminal.UI.Input;
 using YukimaruGames.Terminal.UI.Launcher;
